feat: add <Decoration /> to squares with entities and custom <EntityTooltip />.
Reason (Decoration): In the future, I will add custom .svgs on top of entity cells, which is not possible to full extent with css.
Reason (EntityTooltip): react-tooltips was a havoc to work with. They were either displaying when not needed or not displaying at all (which happened after <Decorator /> introduction.