feat: Invites and Approvals!
This update brings long awaited invitations. Now, you can create invite codes for your lobby and give out for your players! However, until you accept them, they will not be able to access your lobby, so you are secure!

This update utilizes react-query to its fullest and it was a blast using it!